// server/src/lib/logger.d.ts
export declare const logger: {
    debug: (fields: Record<string, unknown>, msg: string) => void;
    info: (fields: Record<string, unknown>, msg: string) => void;
    warn: (fields: Record<string, unknown>, msg: string) => void;
    error: (fields: Record<string, unknown>, msg: string) => void;
};
//# sourceMappingURL=logger.d.ts.map